fix: extract summary to doc summary class
